+++ tests/Type/Validation/ScalarValidationTest.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace GraphQlPhpValidationToolkit\Tests\Type\ValidatedFieldDefinition;
+namespace GraphQlPhpValidationToolkit\Tests\Type\Validation;
 
 use GraphQL\GraphQL;
 use GraphQL\Type\Definition\ObjectType;
